refactor: EntityManagerFactory 추상화
